+++ src/de/interoberlin/sauvignon/model/svg/attributes/SVGTransformTranslate.java
@@ -5,25 +5,25 @@ import de.interoberlin.sauvignon.model.util.Matrix;
 public class SVGTransformTranslate extends ATransformOperator
 {
 	public final ETransformOperators type = ETransformOperators.TRANSLATE;
-	private Double tx = 0d;
-	private Double ty = 0d;
+	private Float tx = 0f;
+	private Float ty = 0f;
 	
 	public SVGTransformTranslate() {}
 	
-	public SVGTransformTranslate(Double t)
+	public SVGTransformTranslate(Float t)
 	{
 		tx = t;
 		updateMatrix = true;
 	}
 	
-	public SVGTransformTranslate(Double tx, Double ty)
+	public SVGTransformTranslate(Float tx, Float ty)
 	{
 		this.tx = tx;
 		this.ty = ty;
 		updateMatrix = true;
 	}
 	
-	public SVGTransformTranslate(Double[] args)
+	public SVGTransformTranslate(Float[] args)
 	{
 		if (args.length > 0)
 		{
@@ -34,23 +34,23 @@ public class SVGTransformTranslate extends ATransformOperator
 		}
 	}
 	
-	public Double getTx()
+	public Float getTx()
 	{
 		return tx;
 	}
 
-	public void setTx(Double tx)
+	public void setTx(Float tx)
 	{
 		this.tx = tx;
 		updateMatrix = true;
 	}
 
-	public Double getTy()
+	public Float getTy()
 	{
 		return ty;
 	}
 
-	public void setTy(Double ty)
+	public void setTy(Float ty)
 	{
 		this.ty = ty;
 		updateMatrix = true;
@@ -60,7 +60,7 @@ public class SVGTransformTranslate extends ATransformOperator
 	{
 		if (updateMatrix)
 		{
-			resultingMatrix = new Matrix(1d, 0d, 0d, 1d, tx, ty);
+			resultingMatrix = new Matrix(1f, 0f, 0f, 1f, tx, ty);
 			updateMatrix = false;
 		}
 		return resultingMatrix;
